i#6662 func_id_filter: large function IDs
When filtering function related markers by function ID we need to
provide the original marker value (i.e., function ID) in the trace.
When filtering out system calls, the ID can be large due to
TRACE_FUNC_ID_SYSCALL_BASE (which is 0x100000000ULL == 4294967296 for x64)
that is added to the system call number, and can cause an std::out_of_range
error.
We now use std::stoll to handle large function IDs.

Issue #6662

+++ clients/drcachesim/tools/filter/record_filter.cpp
@@ -95,7 +95,7 @@ parse_string(const std::string &s, char sep = ',')
     std::vector<T> vec;
     do {
         pos = s.find(sep, at);
-        vec.push_back(static_cast<T>(std::stoi(s.substr(at, pos))));
+        vec.push_back(static_cast<T>(std::stoll(s.substr(at, pos))));
         at = pos + 1;
     } while (pos != std::string::npos);
     return vec;
